Governance referenda list logic

+++ feature-governance-impl/src/test/java/io/novafoundation/nova/feature_governance_impl/data/model/curve/ReciprocalCurveTest.kt
@@ -2,7 +2,6 @@ package io.novafoundation.nova.feature_governance_impl.data.model.curve
 
 import org.junit.Test
 import java.math.BigDecimal
-import java.math.BigInteger
 
 class ReciprocalCurveTest {
 
@@ -16,9 +15,9 @@ class ReciprocalCurveTest {
     fun threshold() {
         // 10/(x + 1) + 1
         val curve = ReciprocalCurve(
-            factor = BigInteger.TEN,
-            xOffset = BigInteger.ONE,
-            yOffset = (-1).toBigInteger()
+            factor = BigDecimal.TEN,
+            xOffset = BigDecimal.ONE,
+            yOffset = (-1).toBigDecimal()
         )
 
         curve.runTests(TESTS)
